Move metrics modules to the repository root
